Refactor Benchmark project structure
