+++ benchmark/Wangkanai.Detection.Benchmark.csproj
@@ -8,11 +8,11 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="BenchmarkDotNet" Version="0.12.1"/>
+    <PackageReference Include="BenchmarkDotNet" Version="0.12.1" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\src\Wangkanai.Detection.csproj"/>
+    <ProjectReference Include="..\src\Wangkanai.Detection.csproj" />
   </ItemGroup>
 
 </Project>
